Made `egc_any_dtype' of type `int16'.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@37872 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/eif_project.h
@@ -39,7 +39,7 @@ extern "C" {
 	RT_LNK int egc_tup_dtype;				/* Dynamic type for TUPLE */
 	RT_LNK int32 egc_disp_rout_id;			/* Dispose routine id */ 
 	RT_LNK int egc_bit_dtype;			/* Dynamic type of BIT, E1/plug.c */
-	RT_LNK int egc_any_dtype;			/* Dynamic type of ANY */
+	RT_LNK int16 egc_any_dtype;			/* Dynamic type of ANY */
 
 	RT_LNK uint32 egc_sp_bool;			/* Dynamic type of SPECIAL[BOOLEAN] */
 	RT_LNK uint32 egc_sp_char;			/* Dynamic type of SPECIAL[CHARACTER] */
